flask_rest/requirements.txt
psycopg2-binary
transformers
torch
flask
gunicorn